move panels to their own file
