+++ packages/keystatic/src/app/shell/index.tsx
@@ -1,19 +1,10 @@
-import { clamp, useLayoutEffect } from '@react-aria/utils';
 import {
   createContext,
   PropsWithChildren,
   ReactElement,
   ReactNode,
   useContext,
-  useRef,
-  useState,
 } from 'react';
-import {
-  PanelGroup,
-  Panel,
-  PanelResizeHandle,
-  ImperativePanelHandle,
-} from 'react-resizable-panels';
 
 import { alertCircleIcon } from '@keystar/ui/icon/icons/alertCircleIcon';
 import { Icon } from '@keystar/ui/icon';
@@ -22,8 +13,6 @@ import {
   VoussoirTheme,
   breakpointQueries,
   css,
-  tokenSchema,
-  transition,
   useMediaQuery,
 } from '@keystar/ui/style';
 import { Heading, Text } from '@keystar/ui/typography';
@@ -39,13 +28,9 @@ import {
   AppShellErrorContext,
   LocalAppShellProvider,
 } from './data';
-import {
-  SidebarDialog,
-  SidebarPanel,
-  SidebarProvider,
-  useSidebar,
-} from './sidebar';
+import { SidebarDialog, SidebarPanel, SidebarProvider } from './sidebar';
 import { TopBar } from './topbar';
+import { MainPanelLayout } from './panels';
 
 export const AppShell = (props: {
   config: Config;
@@ -85,10 +70,10 @@ export const AppShell = (props: {
               {content}
             </>
           ) : (
-            <PanelLayout>
+            <MainPanelLayout>
               <SidebarPanel hrefBase={props.basePath} config={props.config} />
               {content}
-            </PanelLayout>
+            </MainPanelLayout>
           )}
         </Flex>
       </SidebarProvider>
@@ -112,190 +97,6 @@ export const AppShell = (props: {
     );
   }
   return null;
-};
-
-// Panels
-// -----------------------------------------------------------------------------
-
-const SIDEBAR_MIN_PERCENT = 10;
-const SIDEBAR_DEFAULT_PERCENT = 18;
-const SIDEBAR_MAX_PERCENT = 50;
-const SIDEBAR_MIN_PX = 180;
-const SIDEBAR_DEFAULT_PX = 260;
-const SIDEBAR_MAX_PX = 600;
-
-const calcDefault = (t: number) => (SIDEBAR_DEFAULT_PX / t) * 100;
-const calcMin = (t: number) => (SIDEBAR_MIN_PX / t) * 100;
-const calcMax = (t: number) =>
-  Math.min((SIDEBAR_MAX_PX / t) * 100, SIDEBAR_MAX_PERCENT);
-
-function getInitialSizes() {
-  if (typeof window === 'undefined') {
-    return {
-      minSize: SIDEBAR_MIN_PERCENT,
-      maxSize: SIDEBAR_MAX_PERCENT,
-      defaultSize: SIDEBAR_DEFAULT_PERCENT,
-    };
-  }
-
-  // Fallback to `window.innerWidth`, which doesn't include scrollbars but it's
-  // okay for this approximation.
-  let viewportWidth = window.visualViewport?.width || window.innerWidth;
-
-  let minSize = calcMin(viewportWidth);
-  let maxSize = calcMax(viewportWidth);
-  let defaultSize = calcDefault(viewportWidth);
-  return { minSize, maxSize, defaultSize };
-}
-
-const PanelLayout = ({
-  children,
-}: {
-  children: [ReactElement, ReactElement];
-}) => {
-  let [isDragging, setIsDragging] = useState(false);
-  let [size, setSize] = useState(() => getInitialSizes());
-  let sidebarState = useSidebar();
-  let sidebarPanelRef = useRef<ImperativePanelHandle>(null);
-  let [sidebar, content] = children;
-
-  // Sync sidebar context with panel state.
-  useLayoutEffect(() => {
-    let panel = sidebarPanelRef.current;
-    if (panel) {
-      if (sidebarState.isOpen) {
-        panel.resize(size.defaultSize);
-      } else {
-        panel.collapse();
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [sidebarState.isOpen]);
-
-  // Handle cases where the sidebar has an invalid size. This can happen when
-  // the panel has been resized in a larger window, then the window is resized
-  // to be smaller, or vice versa.
-  useLayoutEffect(() => {
-    let panel = sidebarPanelRef.current;
-    if (panel && !panel.getCollapsed()) {
-      let currentSize = panel.getSize();
-      if (currentSize < size.minSize || currentSize > size.maxSize) {
-        panel.resize(clamp(currentSize, size.minSize, size.maxSize));
-      }
-    }
-  }, [size.maxSize, size.minSize]);
-
-  useLayoutEffect(() => {
-    const panelGroup: HTMLElement | null = document.querySelector(
-      '[data-panel-group-id="main"]'
-    );
-    const resizeHandles: NodeListOf<HTMLElement> = document.querySelectorAll(
-      '[data-panel-resize-handle-id]'
-    );
-
-    if (!panelGroup) {
-      return;
-    }
-
-    const observer = new ResizeObserver(() => {
-      let width = panelGroup.offsetWidth;
-
-      // subtract the width of the resize handles
-      resizeHandles.forEach(resizeHandle => {
-        width -= resizeHandle.offsetWidth;
-      });
-
-      let minSize = calcMin(width);
-      let maxSize = calcMax(width);
-      let defaultSize = calcDefault(width);
-      setSize({ minSize, maxSize, defaultSize });
-    });
-    observer.observe(panelGroup);
-    resizeHandles.forEach(resizeHandle => {
-      observer.observe(resizeHandle);
-    });
-
-    return () => {
-      observer.disconnect();
-    };
-  }, []);
-
-  console.log('panels');
-
-  return (
-    <PanelGroup
-      disablePointerEventsDuringResize
-      id="main"
-      autoSaveId="main"
-      direction="horizontal"
-      className={css({ flex: 1 })}
-    >
-      <Panel
-        collapsible
-        defaultSize={size.defaultSize}
-        maxSize={size.maxSize}
-        minSize={size.minSize}
-        onCollapse={isCollapsed => sidebarState.setOpen(!isCollapsed)}
-        ref={sidebarPanelRef}
-        className={css({
-          // containerName: 'sidepanel',
-          // containerType: 'inline-size',
-        })}
-      >
-        {sidebar}
-      </Panel>
-      <PanelResizeHandle
-        onDragging={setIsDragging}
-        disabled={!isDragging && !sidebarState.isOpen}
-        className={css({
-          borderInlineEnd: `1px solid ${tokenSchema.color.border.muted}`,
-          boxSizing: 'border-box',
-          outline: 0,
-          position: 'relative',
-          zIndex: 1,
-
-          // hide when disabled
-          '&[data-panel-resize-handle-enabled=false]': {
-            visibility: 'hidden',
-          },
-
-          // increase hit area
-          '&::before': {
-            content: '""',
-            position: 'absolute',
-            inset: `calc(${tokenSchema.size.space.small} * -1)`,
-          },
-
-          // drag indicator
-          '&::after': {
-            backgroundColor: tokenSchema.color.border.emphasis,
-            content: '""',
-            insetBlock: 0,
-            insetInline: `calc(${tokenSchema.size.space.xsmall} / -2)`,
-            opacity: 0,
-            position: 'absolute',
-            transition: transition('opacity'),
-          },
-          '&:hover::after': {
-            opacity: 1,
-            transition: transition('opacity', { delay: 300 }), // delay to avoid flicker. user may just be mousing around the screen; wait for intent
-          },
-          '&[data-resize-handle-active]::after': {
-            backgroundColor: tokenSchema.color.background.accentEmphasis,
-            opacity: 1,
-          },
-        })}
-      />
-      <Panel
-        className={css({
-          // containerName: 'mainpanel',
-          // containerType: 'inline-size',
-        })}
-      >
-        {content}
-      </Panel>
-    </PanelGroup>
-  );
 };
 
 // Styled components
